Step 1 - break asset group suggest out to component
